Fix and test channel-wise linear normalization
1. Make it actually channel-wise;
2. Before it should take `(old_min, old_max)` of the input image (otherwise it doesn't make sense), but now it takes `(new_min, new_max)` of the expected normalized output.

Before this fix, if `lowhigh` is provided, channel-wise linear normalization is performed by:

```python
for c in ...:
    img_norm[...,c] = (img_norm[..., c] - lowhigh[0]) / (lowhigh[1] - lowhigh[0])
```

which has a redundant `[..., c]` and is not "channel-wise", i.e. it is equivalent to

```python
img_norm = (img_norm - lowhigh[0]) / (lowhigh[1] - lowhigh[0])
```

This parameter, `lowhigh`, is never used in Cellpose itself, however.